fix: checkbox default value was not being accepted properly
When using the model prop on the AvForm, you should be able to pass in an array of names for the components you want to be checked by default in the AvCheckboxGroup. However, the valueArr in this case when using the filter method returns an array of size 0. Instead we can just use the lodash/find property to make sure that if the item is not found that it returns null